feat: 1

+++ src/SaveCacheJob.php
@@ -59,7 +59,7 @@ class SaveCacheJob implements ShouldQueue
                         $cache_key = ModelCache::getStaticCacheKey($this->className, $this->id, $key);
                         if (Cache::has($cache_key)) {
                             $value      = Cache::pull($cache_key);
-                            $cacheValue = (int)bcdiv($value, 1000, 5);
+                            $cacheValue = (float)bcdiv($value, 1000, 5);
 //                    info("缓存Key: $cache_key ,值: $cacheValue 模型值: {$model->{$key}}");
                             if ($model->{$key} != $cacheValue) {
                                 $model->{$key} = $cacheValue;
